hidpp20: Move profile report rate decoding/encoding
This change moves where the value of the profiles report rate is
decoded/encoded from the parse_profile/write_profile functions to the
read_profile_8100/update_report_rate_8100 functions.

This makes report rate consistent with other internal profile values,
such that the raw value is stored in the internal profile structure.
Additionally this refactors the internal profile and decoding/encoding
handling to integrate better with additional profile formats or report
rate feature capabilities (e.g. 0x8061).

+++ src/hidpp20.h
@@ -952,7 +952,7 @@ struct hidpp20_profile {
 	char name[16 * 3];
 	uint16_t powersave_timeout;
 	uint16_t poweroff_timeout;
-	unsigned report_rate;
+	uint8_t report_rate;
 	unsigned default_dpi;
 	unsigned switched_dpi;
 	unsigned current_dpi;
